removed commit paths & moved some logic to zerostash-fuse & fuse feature flag
