+++ zerostash-fuse/Cargo.toml
@@ -1,0 +1,10 @@
+[package]
+name = "zerostash-fuse"
+version = "0.1.0"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+infinitree = { version = "0.9.0", features = ["cryptobox", "yubikey"] }
+serde = { version = "1.0.145" }
