+++ zerostash/Cargo.toml
@@ -17,6 +17,7 @@ clap = "4.2.5"
 infinitree = { version = "0.10.0", features = ["cryptobox", "yubikey"] }
 infinitree-backends = { version = "0.10.0", default-features = false, features = ["rustls"] }
 zerostash-files = { version = "0.6.0", path = "../zerostash-files" }
+zerostash-fuse = { version = "0.1.0", path = "../zerostash-fuse", optional = true}
 rpassword = "7.2.0"
 rprompt = "2.0.2"
 serde = { version = "1.0.160", features = ["serde_derive"] }
@@ -40,6 +41,13 @@ ctrlc = "3.2.3"
 tracing = "0.1.37"
 
 secrecy = { version = "0.8.0", features = ["serde"] }
+scc = {version = "1.1.0", optional = true}
+fuse_mt = {version = "0.6.0", optional = true}
+tracing = {version = "0.1.37", optional = true}
+ctrlc = {version = "3.2.3", optional = true}
+
+[features]
+fuse = ["dep:zerostash-fuse", "dep:scc", "dep:fuse_mt", "dep:ctrlc", "dep:tracing"]
 
 [target.'cfg(target_os = "macos")'.dependencies]
 security-framework = "2.8.2"
@@ -53,13 +61,12 @@ tracing-subscriber = "0.3.17"
 tracing = "0.1.37"
 criterion = "0.3.6"
 
+[[bench]]
+name = "bench"
+harness = false
+path = "benches/bench.rs"
 
 #[[bench]]
 #name = "bench"
 #harness = false
 #path = "../bench/main.rs"
-
-[[bench]]
-name = "bench-cri"
-harness = false
-path = "benches/bench.rs"
