+++ zerostash/src/commands.rs
@@ -14,8 +14,10 @@ mod wipe;
 use wipe::*;
 mod zfs;
 use zfs::*;
-mod mount;
-use mount::*;
+#[cfg(feature = "fuse")]
+pub mod mount;
+#[cfg(feature = "fuse")]
+pub use mount::*;
 
 use crate::{
     config::{Key, SymmetricKey, YubikeyCRConfig, YubikeyCRKey},
@@ -46,6 +48,7 @@ pub enum ZerostashCmd {
     Ls(Ls),
 
     /// Mount the files in a stash
+    #[cfg(feature = "fuse")]
     Mount(Mount),
 
     /// Key management & generation
@@ -163,9 +166,10 @@ impl Runnable for EntryPoint {
                 Log(cmd) => cmd.run().await,
                 Ls(cmd) => cmd.run().await,
                 Keys(cmd) => cmd.run().await,
-                Mount(cmd) => cmd.run().await,
                 Wipe(cmd) => cmd.run().await,
                 Zfs(cmd) => cmd.run().await,
+                #[cfg(feature = "fuse")]
+                Mount(cmd) => cmd.run().await,
             }
         })
         .unwrap()
